Move Model to its own file
